requirements.txt
pdf2image==1.17.*
pillow==10.4.*
reportlab==4.2.*
typer==0.12.*
